feat(autoware_auto_perception_rviz_plugin): rename package to autoware_perception_rviz_plugin
