+++ common/autoware_perception_rviz_plugin/icons/classes/TrackedObjects.png
