+++ common/autoware_perception_rviz_plugin/images/detected-object-visualization-description.jpg
