+++ common/autoware_perception_rviz_plugin/images/predicted-object-visualization-description.jpg
