+++ common/autoware_perception_rviz_plugin/images/tracked-object-visualization-description.jpg
