+++ common/autoware_perception_rviz_plugin/include/autoware_perception_rviz_plugin/object_detection/detected_objects_display.hpp
@@ -11,10 +11,10 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-#ifndef AUTOWARE_AUTO_PERCEPTION_RVIZ_PLUGIN__OBJECT_DETECTION__DETECTED_OBJECTS_DISPLAY_HPP_
-#define AUTOWARE_AUTO_PERCEPTION_RVIZ_PLUGIN__OBJECT_DETECTION__DETECTED_OBJECTS_DISPLAY_HPP_
+#ifndef AUTOWARE_PERCEPTION_RVIZ_PLUGIN__OBJECT_DETECTION__DETECTED_OBJECTS_DISPLAY_HPP_
+#define AUTOWARE_PERCEPTION_RVIZ_PLUGIN__OBJECT_DETECTION__DETECTED_OBJECTS_DISPLAY_HPP_
 
-#include "autoware_auto_perception_rviz_plugin/object_detection/object_polygon_display_base.hpp"
+#include "autoware_perception_rviz_plugin/object_detection/object_polygon_display_base.hpp"
 
 #include <autoware_auto_perception_msgs/msg/detected_objects.hpp>
 
@@ -25,7 +25,7 @@ namespace rviz_plugins
 namespace object_detection
 {
 /// \brief Class defining rviz plugin to visualize DetectedObjects
-class AUTOWARE_AUTO_PERCEPTION_RVIZ_PLUGIN_PUBLIC DetectedObjectsDisplay
+class AUTOWARE_PERCEPTION_RVIZ_PLUGIN_PUBLIC DetectedObjectsDisplay
 : public ObjectPolygonDisplayBase<autoware_auto_perception_msgs::msg::DetectedObjects>
 {
   Q_OBJECT
@@ -43,4 +43,4 @@ private:
 }  // namespace rviz_plugins
 }  // namespace autoware
 
-#endif  // AUTOWARE_AUTO_PERCEPTION_RVIZ_PLUGIN__OBJECT_DETECTION__DETECTED_OBJECTS_DISPLAY_HPP_
+#endif  // AUTOWARE_PERCEPTION_RVIZ_PLUGIN__OBJECT_DETECTION__DETECTED_OBJECTS_DISPLAY_HPP_
