+++ common/autoware_perception_rviz_plugin/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>autoware_auto_perception_rviz_plugin</name>
+  <name>autoware_perception_rviz_plugin</name>
   <version>1.0.0</version>
   <description>Contains plugins to visualize object detection outputs</description>
   <maintainer email="opensource@apex.ai">Apex.AI, Inc.</maintainer>
