+++ common/autoware_perception_rviz_plugin/src/object_detection/tracked_objects_display.cpp
@@ -14,7 +14,7 @@
 //
 // Co-developed by Tier IV, Inc. and Apex.AI, Inc.
 
-#include "autoware_auto_perception_rviz_plugin/object_detection/detected_objects_display.hpp"
+#include "autoware_perception_rviz_plugin/object_detection/tracked_objects_display.hpp"
 
 #include <memory>
 
@@ -24,16 +24,37 @@ namespace rviz_plugins
 {
 namespace object_detection
 {
-DetectedObjectsDisplay::DetectedObjectsDisplay() : ObjectPolygonDisplayBase("detected_objects")
+TrackedObjectsDisplay::TrackedObjectsDisplay() : ObjectPolygonDisplayBase("tracks")
 {
+  // Option for selective visualization by object dynamics
+  m_select_object_dynamics_property = new rviz_common::properties::EnumProperty(
+    "Dynamic Status", "All", "Selectively visualize objects by its dynamic status.", this);
+  m_select_object_dynamics_property->addOption("Dynamic", 0);
+  m_select_object_dynamics_property->addOption("Static", 1);
+  m_select_object_dynamics_property->addOption("All", 2);
 }
 
-void DetectedObjectsDisplay::processMessage(DetectedObjects::ConstSharedPtr msg)
+bool TrackedObjectsDisplay::is_object_to_show(
+  const uint showing_dynamic_status, const TrackedObject & object)
+{
+  if (showing_dynamic_status == 0 && object.kinematics.is_stationary) {
+    return false;  // Show only moving objects
+  }
+  if (showing_dynamic_status == 1 && !object.kinematics.is_stationary) {
+    return false;  // Show only stationary objects
+  }
+  return true;
+}
+
+void TrackedObjectsDisplay::processMessage(TrackedObjects::ConstSharedPtr msg)
 {
   clear_markers();
-  int id = 0;
+  update_id_map(msg);
+
+  const auto showing_dynamic_status = get_object_dynamics_to_visualize();
   for (const auto & object : msg->objects) {
-    // TODO(Satoshi Tanaka): fixing from string label to one string
+    // Filter by object dynamic status
+    if (!is_object_to_show(showing_dynamic_status, object)) continue;
     // Get marker for shape
     auto shape_marker = get_shape_marker_ptr(
       object.shape, object.kinematics.pose_with_covariance.pose.position,
@@ -44,7 +65,7 @@ void DetectedObjectsDisplay::processMessage(DetectedObjects::ConstSharedPtr msg)
     if (shape_marker) {
       auto shape_marker_ptr = shape_marker.value();
       shape_marker_ptr->header = msg->header;
-      shape_marker_ptr->id = id++;
+      shape_marker_ptr->id = uuid_to_marker_id(object.object_id);
       add_marker(shape_marker_ptr);
     }
 
@@ -55,8 +76,23 @@ void DetectedObjectsDisplay::processMessage(DetectedObjects::ConstSharedPtr msg)
     if (label_marker) {
       auto label_marker_ptr = label_marker.value();
       label_marker_ptr->header = msg->header;
-      label_marker_ptr->id = id++;
+      label_marker_ptr->id = uuid_to_marker_id(object.object_id);
       add_marker(label_marker_ptr);
+    }
+
+    // Get marker for id
+    geometry_msgs::msg::Point uuid_vis_position;
+    uuid_vis_position.x = object.kinematics.pose_with_covariance.pose.position.x - 0.5;
+    uuid_vis_position.y = object.kinematics.pose_with_covariance.pose.position.y;
+    uuid_vis_position.z = object.kinematics.pose_with_covariance.pose.position.z - 0.5;
+
+    auto id_marker =
+      get_uuid_marker_ptr(object.object_id, uuid_vis_position, object.classification);
+    if (id_marker) {
+      auto id_marker_ptr = id_marker.value();
+      id_marker_ptr->header = msg->header;
+      id_marker_ptr->id = uuid_to_marker_id(object.object_id);
+      add_marker(id_marker_ptr);
     }
 
     // Get marker for pose covariance
@@ -65,7 +101,7 @@ void DetectedObjectsDisplay::processMessage(DetectedObjects::ConstSharedPtr msg)
     if (pose_with_covariance_marker) {
       auto marker_ptr = pose_with_covariance_marker.value();
       marker_ptr->header = msg->header;
-      marker_ptr->id = id++;
+      marker_ptr->id = uuid_to_marker_id(object.object_id);
       add_marker(marker_ptr);
     }
 
@@ -76,7 +112,7 @@ void DetectedObjectsDisplay::processMessage(DetectedObjects::ConstSharedPtr msg)
     if (yaw_covariance_marker) {
       auto marker_ptr = yaw_covariance_marker.value();
       marker_ptr->header = msg->header;
-      marker_ptr->id = id++;
+      marker_ptr->id = uuid_to_marker_id(object.object_id);
       add_marker(marker_ptr);
     }
 
@@ -92,22 +128,36 @@ void DetectedObjectsDisplay::processMessage(DetectedObjects::ConstSharedPtr msg)
     if (existence_prob_marker) {
       auto existence_prob_marker_ptr = existence_prob_marker.value();
       existence_prob_marker_ptr->header = msg->header;
-      existence_prob_marker_ptr->id = id++;
+      existence_prob_marker_ptr->id = uuid_to_marker_id(object.object_id);
       add_marker(existence_prob_marker_ptr);
     }
-
     // Get marker for velocity text
     geometry_msgs::msg::Point vel_vis_position;
-    vel_vis_position.x = object.kinematics.pose_with_covariance.pose.position.x - 0.5;
-    vel_vis_position.y = object.kinematics.pose_with_covariance.pose.position.y;
-    vel_vis_position.z = object.kinematics.pose_with_covariance.pose.position.z - 0.5;
+    vel_vis_position.x = uuid_vis_position.x - 0.5;
+    vel_vis_position.y = uuid_vis_position.y;
+    vel_vis_position.z = uuid_vis_position.z - 0.5;
     auto velocity_text_marker = get_velocity_text_marker_ptr(
       object.kinematics.twist_with_covariance.twist, vel_vis_position, object.classification);
     if (velocity_text_marker) {
       auto velocity_text_marker_ptr = velocity_text_marker.value();
       velocity_text_marker_ptr->header = msg->header;
-      velocity_text_marker_ptr->id = id++;
+      velocity_text_marker_ptr->id = uuid_to_marker_id(object.object_id);
       add_marker(velocity_text_marker_ptr);
+    }
+
+    // Get marker for acceleration text
+    geometry_msgs::msg::Point acc_vis_position;
+    acc_vis_position.x = uuid_vis_position.x - 1.0;
+    acc_vis_position.y = uuid_vis_position.y;
+    acc_vis_position.z = uuid_vis_position.z - 1.0;
+    auto acceleration_text_marker = get_acceleration_text_marker_ptr(
+      object.kinematics.acceleration_with_covariance.accel, acc_vis_position,
+      object.classification);
+    if (acceleration_text_marker) {
+      auto acceleration_text_marker_ptr = acceleration_text_marker.value();
+      acceleration_text_marker_ptr->header = msg->header;
+      acceleration_text_marker_ptr->id = uuid_to_marker_id(object.object_id);
+      add_marker(acceleration_text_marker_ptr);
     }
 
     // Get marker for twist
@@ -117,7 +167,7 @@ void DetectedObjectsDisplay::processMessage(DetectedObjects::ConstSharedPtr msg)
     if (twist_marker) {
       auto twist_marker_ptr = twist_marker.value();
       twist_marker_ptr->header = msg->header;
-      twist_marker_ptr->id = id++;
+      twist_marker_ptr->id = uuid_to_marker_id(object.object_id);
       add_marker(twist_marker_ptr);
     }
 
@@ -127,7 +177,7 @@ void DetectedObjectsDisplay::processMessage(DetectedObjects::ConstSharedPtr msg)
     if (twist_covariance_marker) {
       auto marker_ptr = twist_covariance_marker.value();
       marker_ptr->header = msg->header;
-      marker_ptr->id = id++;
+      marker_ptr->id = uuid_to_marker_id(object.object_id);
       add_marker(marker_ptr);
     }
 
@@ -138,18 +188,18 @@ void DetectedObjectsDisplay::processMessage(DetectedObjects::ConstSharedPtr msg)
     if (yaw_rate_marker) {
       auto marker_ptr = yaw_rate_marker.value();
       marker_ptr->header = msg->header;
-      marker_ptr->id = id++;
+      marker_ptr->id = uuid_to_marker_id(object.object_id);
       add_marker(marker_ptr);
     }
 
-    // Get marker for yaw rate covariance
+    // Get marker for twist covariance
     auto yaw_rate_covariance_marker = get_yaw_rate_covariance_marker_ptr(
       object.kinematics.pose_with_covariance, object.kinematics.twist_with_covariance,
-      get_line_width() * 0.3);
+      get_line_width() * 0.5);
     if (yaw_rate_covariance_marker) {
       auto marker_ptr = yaw_rate_covariance_marker.value();
       marker_ptr->header = msg->header;
-      marker_ptr->id = id++;
+      marker_ptr->id = uuid_to_marker_id(object.object_id);
       add_marker(marker_ptr);
     }
   }
@@ -162,4 +212,4 @@ void DetectedObjectsDisplay::processMessage(DetectedObjects::ConstSharedPtr msg)
 // Export the plugin
 #include <pluginlib/class_list_macros.hpp>  // NOLINT
 PLUGINLIB_EXPORT_CLASS(
-  autoware::rviz_plugins::object_detection::DetectedObjectsDisplay, rviz_common::Display)
+  autoware::rviz_plugins::object_detection::TrackedObjectsDisplay, rviz_common::Display)
